% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCensusData.R
\name{addCensusData}
\alias{addCensusData}
\title{addCensusData: function for fetching census data and merging it with input dataset}
\usage{
addCensusData(db = dbViewR::selectFromDB(), geography = "tract",
  variables, year, state = "WA", county = "King",
  source = "decennial", credentials_path = "/home/rstudio/seattle_flu")
}
\arguments{
\item{db}{tibble with valid column names for INLA model}

\item{geography}{geography of your data, default value - 'tract', for other options check here}

\item{variables}{character string or vector of character strings of variable IDs.}

\item{year}{the year for which you are requesting data}

\item{state}{the state for which you are requesting data, default value - 'WA'}

\item{county}{the county for which you are requesting data, default value - 'King'}

\item{source}{source database, one of: 'acs' for five-year American Community Survey or 'decennial' for decennial Census}

\item{credentials_path}{path to your file with Census API key, you can get your own census api key here: https://api.census.gov/data/key_signup.html}
}
\value{
db tibble with flu and census data
}
\description{
addCensusData: function for fetching census data and merging it with input dataset
}
\examples{

}
