% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandDB.R
\name{expandDB}
\alias{expandDB}
\title{expandDB: function to expand database to fill in unobserved elements for prediction/interpolation}
\usage{
expandDB(db = dbViewR::selectFromDB(),
  linelist = dbViewR::selectFromDB(db$queryList[names(db$queryList)
  \%in\% c("SELECT", "WHERE")]),
  shp = dbViewR::masterSpatialDB(shape_level = "census_tract", source =
  "king_county_geojson"))
}
\arguments{
\item{db}{tibble with valid column names for INLA model}
}
\value{
db tibble with filled in NaNs or Zeros as required
}
\description{
expandDB: function to expand database to fill in unobserved elements for prediction/interpolation
}
\examples{

}
