% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBingData.R
\name{loadBingData}
\alias{loadBingData}
\title{addCensusData: function for loading bing search queries and categorizing which census tract it's in}
\usage{
loadBingData(geography = "tract",
  bing_queries_path = "/home/rstudio/seattle_flu")
}
\arguments{
\item{geography}{geography to bin the bing queries, default value - 'tract'}

\item{bing_queries_path}{path to your file with Census API key, you can get your own census api key here: https://api.census.gov/data/key_signup.html}
}
\value{
db dataframe with number of bing queries per tract
}
\description{
addCensusData: function for loading bing search queries and categorizing which census tract it's in
}
\examples{

}
