% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFromDB.R
\name{masterSpatialDB}
\alias{masterSpatialDB}
\title{masterSpatialDB: function for fetching spatial data from master source}
\usage{
masterSpatialDB(shape_level = "census_tract",
  source = "seattle_geojson", rm_files = TRUE)
}
\arguments{
\item{shape_level}{one of "census_tract" (default),"cra_name","neighborhood","city"}

\item{source}{source database, one of "seattle_geojson" (default), "simulated_data"}

\item{rm_files}{indicator to remove local files (TRUE == default)}
}
\value{
sf object with shapefile data
}
\description{
masterSpatialDB: function for fetching spatial data from master source
}
\examples{
   shp <- masterSpatialDB(shape_level = 'census_tract', source = 'seattle_geojson', rm_files = TRUE)

}
