% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFromDB.R
\name{selectFromDB}
\alias{selectFromDB}
\title{selectFromDB: function for fetching data from research DB
(Currently pulls from simulated data at https://github.com/seattleflu/simulated-data)}
\usage{
selectFromDB(queryIn = jsonlite::toJSON(list(SELECT = list(COLUMN =
  c("pathogen", "encountered_date", "residence_puma5ce",
  "residence_census_tract")), MUTATE = list(COLUMN = c("encountered_date"),
  AS = "encountered_week"), GROUP_BY = list(COLUMN = c("encountered_week",
  "residence_puma5ce", "residence_census_tract")), SUMMARIZE = list(COLUMN
  = "pathogen", IN = c("h1n1pdm")))), source = "simulated_data",
  credentials_path = "/home/rstudio/seattle_flu", na.rm = FALSE)
}
\arguments{
\item{queryIn}{list or json specifying query  (See example)}

\item{source}{source database, one of: 'simulated_data' (default) or 'production'}

\item{credentials_path}{path to your pg_service and pgpass file for production database}

\item{na.rm}{= FALSE (default) Drop rows with NA from dataset as incidenceMapR will ignore them anyway}
}
\value{
dbViewR list with query and observedData table that has been prepared for defineModels.R
}
\description{
STANDARD DB QUERIES WILL ALL LIKELY MIGRATE TO THE HUTCH RESARCH DB BEFORE PRODUCTION.
}
\examples{
return h1n1pdm summary by time and location
queryJSON <- jsonlite::toJSON(
  list(
      SELECT   =list(COLUMN=c('pathogen','encountered_date','residence_puma5ce','residence_census_tract')),
      MUTATE   =list(COLUMN=c('encountered_date'), AS=c('epi_week')),
      GROUP_BY =list(COLUMN=c('epi_week','residence_puma5ce','residence_census_tract')),
      SUMMARIZE=list(COLUMN='pathogen', IN= c('h1n1pdm'))
      )
   )
   db <- selectFromDB( queryJSON )

}
