% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFromDB.R
\name{masterSpatialDB}
\alias{masterSpatialDB}
\title{masterSpatialDB: function for fetching spatial data from master source
(Currently pulls from https://github.com/seattleflu/simulated-data/tree/master/kingCountySpatialData)}
\usage{
masterSpatialDB()
}
\value{
sf object with king county shapefile data
}
\description{
masterSpatialDB: function for fetching spatial data from master source
(Currently pulls from https://github.com/seattleflu/simulated-data/tree/master/kingCountySpatialData)
}
\examples{
   shp <- masterSpatialDB()

}
