% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFromDB.R
\name{selectFromDB}
\alias{selectFromDB}
\title{selectFromDB: function for fetching data from research DB
(Currently pulls from simulated data at https://github.com/seattleflu/simulated-data)}
\usage{
selectFromDB(queryIn = jsonlite::toJSON(list(SELECT = list(COLUMN =
  c("pathogen", "num_date", "PUMA5CE", "GEOID")), MUTATE = list(COLUMN =
  c("num_date"), AS = c("timeBin")), GROUP_BY = list(COLUMN = c("timeBin",
  "PUMA5CE", "GEOID")), SUMMARIZE = list(COLUMN = "pathogen", IN =
  c("h1n1pdm")))))
}
\arguments{
\item{queryIn}{list or json specifying query  (See example)}
}
\value{
observedData table that has been prepared for defineModels.R
}
\description{
STANDARD DB QUERIES WILL ALL LIKELY MIGRATE TO THE HUTCH RESARCH DB BEFORE PRODUCTION.
}
\examples{
return h1n1pdm summary by time and location
queryJSON <- jsonlite::toJSON(
  list(
      SELECT   =list(COLUMN=c('pathogen','num_date','PUMA5CE','GEOID')),
      MUTATE   =list(COLUMN=c('num_date'), AS=c('timeBin')),
      GROUP_BY =list(COLUMN=c('timeBin','PUMA5CE','GEOID')),
      SUMMARIZE=list(COLUMN='pathogen', IN= c('h1n1pdm'))
      )
   )
   db <- selectFromDB( queryJSON )

}
