% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFromDB.R
\name{selectFromDB}
\alias{selectFromDB}
\title{selectFromDB: function for fetching data from research DB
(Currently pulls from simulated data at https://github.com/seattleflu/simulated-data)}
\usage{
selectFromDB(queryIn = jsonlite::toJSON(list(SELECT = list(COLUMN =
  c("pathogen", "encountered_date", "residence_puma",
  "residence_census_tract")), GROUP_BY = list(COLUMN =
  c("encountered_week", "residence_puma", "residence_census_tract")),
  SUMMARIZE = list(COLUMN = "pathogen", IN = c("h1n1pdm")))),
  source = "simulated_data",
  credentials_path = "/home/rstudio/seattle_flu", na.rm = FALSE,
  shp = dbViewR::masterSpatialDB(shape_level = "census_tract", source =
  "king_county_geojson"))
}
\arguments{
\item{queryIn}{list or json specifying query  (See example)}

\item{source}{source database, one of: 'simulated_data' (default) or 'production'}

\item{credentials_path}{path to your pg_service and pgpass file for production database}

\item{na.rm}{= FALSE (default) Drop rows with NA from dataset as incidenceMapR will ignore them anyway}

\item{shp}{= dbViewR::masterSpatialDB(shape_level = 'census_tract', source = 'king_county_geojson') (Needed hack until higher-level shape labels are in database)}
}
\value{
dbViewR list with query and observedData table that has been prepared for defineModels.R
}
\description{
STANDARD DB QUERIES WILL ALL LIKELY MIGRATE TO THE HUTCH RESARCH DB BEFORE PRODUCTION.
}
\examples{
return h1n1pdm summary by time and location
queryJSON <- jsonlite::toJSON(
  list(
      SELECT   =list(COLUMN=c('pathogen','encountered_week','residence_puma','residence_census_tract')),
      GROUP_BY =list(COLUMN=c('encountered_week','residence_puma','residence_census_tract')),
      SUMMARIZE=list(COLUMN='pathogen', IN= c('h1n1pdm'))
      )
   )
   db <- selectFromDB( queryJSON )

}
